+++ HelpScoutNet.ConsoleClient/HelpScoutNet.ConsoleClient.csproj
@@ -41,6 +41,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Models\AssignedTo.cs" />
+    <Compile Include="Models\CreatedBy.cs" />
+    <Compile Include="Models\Customer.cs" />
+    <Compile Include="Models\Item.cs" />
+    <Compile Include="Models\Mailbox.cs" />
+    <Compile Include="Models\Owner.cs" />
+    <Compile Include="Models\Source.cs" />
+    <Compile Include="Models\Thread.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
