+++ HelpScoutNet.ConsoleClient/Models/AssignedTo.cs
@@ -1,0 +1,12 @@
+﻿namespace HelpScoutNet.ConsoleClient.Models
+{
+    public class AssignedTo
+    {
+        public int id { get; set; }
+        public string firstName { get; set; }
+        public string lastName { get; set; }
+        public string email { get; set; }
+        public object phone { get; set; }
+        public string type { get; set; }
+    }
+}
