+++ HelpScoutNet.ConsoleClient/Models/Mailbox.cs
@@ -1,0 +1,8 @@
+﻿namespace HelpScoutNet.ConsoleClient.Models
+{
+    public class Mailbox
+    {
+        public int id { get; set; }
+        public string name { get; set; }
+    }
+}
