+++ HelpScoutNet.ConsoleClient/Models/Source.cs
@@ -1,0 +1,8 @@
+﻿namespace HelpScoutNet.ConsoleClient.Models
+{
+    public class Source
+    {
+        public string type { get; set; }
+        public string via { get; set; }
+    }
+}
